refactor: extract DiagnosticsTypes + AppLaunchDiagnosticsReporter
Bundles 2 self-contained extractions from BugNarratorDiagnostics.swift:

**DiagnosticsTypes.swift** (89 lines):
- enum DiagnosticsLogCategory
- enum DiagnosticsLogLevel
- struct DiagnosticsEventName (RawRepresentable, ExpressibleByStringLiteral)
- enum DiagnosticsEvent
- extension DiagnosticsEventName
- SHA-256: ade6ef00768df93561c21559c957d4501e3959d5f044c4c720ed3dc5b89b241f
  (lines 4-92 of origin/main:BugNarratorDiagnostics.swift, verified)

**AppLaunchDiagnosticsReporter.swift** (29 lines):
- @MainActor final class AppLaunchDiagnosticsReporter
- SHA-256: 96029cbf050515c7889f1da7f1a225c32685c3e70d1059892e562c03f03d11a3
  (lines 371-398 of origin/main, verified)

Applied memory rule "watch for annotations on line immediately preceding
deleted blocks": verified @MainActor at line 370 is INSIDE the deleted range
(belongs to the extracted AppLaunchDiagnosticsReporter class) — it's the
CORRECT case where the annotation must move with the block.

Imports added: OSLog (for OSLogType used in DiagnosticsLogLevel).

Reduces BugNarratorDiagnostics.swift from 571 to ~452 lines.

Tests: 667 (unchanged).
